Cleanup demo app (#265)
