+++ package.json
@@ -71,9 +71,9 @@
         "@maplibre/maplibre-gl-style-spec": "^19.3.3",
         "@open-pioneer/chakra-integration": "^1.1.1",
         "@open-pioneer/core": "^1.2.1",
-        "@open-pioneer/integration": "^2.0.2",
-        "@open-pioneer/http": "^2.1.1",
-        "@open-pioneer/runtime": "^2.0.2",
+        "@open-pioneer/http": "^2.1.2",
+        "@open-pioneer/integration": "^2.0.3",
+        "@open-pioneer/runtime": "^2.1.0",
         "@open-pioneer/test-utils": "^1.0.2",
         "framer-motion": "^10.17.4",
         "ol": "^8.2.0",
