+++ src/experimental-packages/layout-sidebar/package.json
@@ -12,7 +12,7 @@
         "@emotion/react": "^11.11.3",
         "@emotion/styled": "^11.11.0",
         "@open-pioneer/chakra-integration": "^1.1.1",
-        "@open-pioneer/runtime": "^2.0.2",
+        "@open-pioneer/runtime": "^2.1.0",
         "react": "^18.2.0"
     },
     "devDependencies": {
