+++ src/packages/authentication/package.json
@@ -9,7 +9,7 @@
     },
     "dependencies": {},
     "peerDependencies": {
-        "@open-pioneer/runtime": "^2.0.2",
+        "@open-pioneer/runtime": "^2.1.0",
         "@open-pioneer/core": "^1.2.1",
         "react": "^18.2.0",
         "react-use": "^17.4.2"
