+++ src/packages/geolocation/package.json
@@ -9,7 +9,7 @@
     "peerDependencies": {
         "@open-pioneer/chakra-integration": "^1.1.1",
         "@open-pioneer/map": "workspace:^",
-        "@open-pioneer/runtime": "^2.0.2",
+        "@open-pioneer/runtime": "^2.1.0",
         "@open-pioneer/react-utils": "workspace:^",
         "@open-pioneer/core": "^1.2.1",
         "@open-pioneer/notifier": "workspace:^",
