+++ src/packages/local-storage/package.json
@@ -8,7 +8,7 @@
     },
     "peerDependencies": {
         "@open-pioneer/core": "^1.2.1",
-        "@open-pioneer/runtime": "^2.0.2"
+        "@open-pioneer/runtime": "^2.1.0"
     },
     "devDependencies": {
         "@open-pioneer/test-utils": "^1.0.2",
