+++ src/packages/ogc-features/package.json
@@ -8,7 +8,7 @@
     },
     "peerDependencies": {
         "@open-pioneer/core": "^1.2.1",
-        "@open-pioneer/http": "^2.1.1",
+        "@open-pioneer/http": "^2.1.2",
         "ol": "^8.2.0",
         "uuid": "^9.0.1",
         "@open-pioneer/search": "workspace:^"
