+++ src/packages/overview-map/package.json
@@ -10,7 +10,7 @@
         "@open-pioneer/chakra-integration": "^1.1.1",
         "@open-pioneer/map": "workspace:^",
         "@open-pioneer/ogc-features": "workspace:^",
-        "@open-pioneer/runtime": "^2.0.2",
+        "@open-pioneer/runtime": "^2.1.0",
         "@open-pioneer/react-utils": "workspace:^",
         "ol": "^8.2.0",
         "react": "^18.2.0"
