+++ src/packages/spatial-bookmarks/package.json
@@ -13,7 +13,7 @@
         "@open-pioneer/local-storage": "workspace:^",
         "@open-pioneer/map": "workspace:^",
         "@open-pioneer/react-utils": "workspace:^",
-        "@open-pioneer/runtime": "^2.0.2",
+        "@open-pioneer/runtime": "^2.1.0",
         "classnames": "^2.3.2",
         "ol": "^8.2.0",
         "react": "^18.2.0",
