+++ src/packages/toc/package.json
@@ -11,7 +11,7 @@
         "@open-pioneer/map": "workspace:^",
         "@open-pioneer/basemap-switcher": "workspace:^",
         "@open-pioneer/react-utils": "workspace:^",
-        "@open-pioneer/runtime": "^2.0.2",
+        "@open-pioneer/runtime": "^2.1.0",
         "ol": "^8.2.0",
         "react": "^18.2.0",
         "classnames": "^2.3.2",
