+++ src/samples/auth-sample/auth-app/package.json
@@ -6,7 +6,7 @@
         "@open-pioneer/authentication": "workspace:^",
         "@open-pioneer/chakra-integration": "^1.1.1",
         "@open-pioneer/core": "^1.2.1",
-        "@open-pioneer/runtime": "^2.0.2",
+        "@open-pioneer/runtime": "^2.1.0",
         "@open-pioneer/theme": "workspace:^",
         "react": "^18.2.0"
     }
