+++ src/samples/experimental-sidebar/sidebar-app/package.json
@@ -11,7 +11,7 @@
         "@open-pioneer/experimental-layout-sidebar": "workspace:^",
         "@open-pioneer/map": "workspace:^",
         "@open-pioneer/react-utils": "workspace:^",
-        "@open-pioneer/runtime": "^2.0.2",
+        "@open-pioneer/runtime": "^2.1.0",
         "@open-pioneer/theme": "workspace:^",
         "ol": "^8.2.0",
         "react": "^18.2.0",
