--- src/samples/map-sample/map-sample-logging/package.json
@@ -1,7 +1,0 @@
-{
-    "name": "map-sample-logging",
-    "private": true,
-    "devDependencies": {
-        "@open-pioneer/runtime": "^2.0.2"
-    }
-}
