+++ src/samples/map-sample/ol-app/package.json
@@ -7,27 +7,26 @@
         "@emotion/react": "^11.11.3",
         "@emotion/styled": "^11.11.0",
         "@open-pioneer/chakra-integration": "^1.1.1",
-        "@open-pioneer/http": "^2.1.1",
         "@open-pioneer/coordinate-viewer": "workspace:^",
-        "@open-pioneer/selection": "workspace:^",
         "@open-pioneer/experimental-layout-sidebar": "workspace:^",
         "@open-pioneer/geolocation": "workspace:^",
+        "@open-pioneer/http": "^2.1.2",
+        "@open-pioneer/legend": "workspace:^",
         "@open-pioneer/map": "workspace:^",
-        "@open-pioneer/ogc-features": "workspace:^",
         "@open-pioneer/map-navigation": "workspace:^",
         "@open-pioneer/measurement": "workspace:^",
         "@open-pioneer/notifier": "workspace:^",
+        "@open-pioneer/ogc-features": "workspace:^",
         "@open-pioneer/overview-map": "workspace:^",
         "@open-pioneer/react-utils": "workspace:^",
-        "@open-pioneer/runtime": "^2.0.2",
+        "@open-pioneer/runtime": "^2.1.0",
         "@open-pioneer/scale-bar": "workspace:^",
         "@open-pioneer/scale-viewer": "workspace:^",
         "@open-pioneer/search": "workspace:^",
+        "@open-pioneer/selection": "workspace:^",
         "@open-pioneer/spatial-bookmarks": "workspace:^",
         "@open-pioneer/theme": "workspace:^",
         "@open-pioneer/toc": "workspace:^",
-        "map-sample-logging": "workspace:^",
-        "@open-pioneer/legend": "workspace:^",
         "ol": "^8.2.0",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
