+++ src/samples/map-sample/ol-app/services.ts
@@ -1,4 +1,4 @@
 // SPDX-FileCopyrightText: 2023 Open Pioneer project (https://github.com/open-pioneer)
 // SPDX-License-Identifier: Apache-2.0
 export { MapConfigProviderImpl } from "./MapConfigProviderImpl";
-export { AppConfig } from "./AppConfig";
+export { AppModel } from "./AppModel";
