+++ src/samples/notify-sample/notify-app/package.json
@@ -5,7 +5,7 @@
     "dependencies": {
         "@open-pioneer/chakra-integration": "^1.1.1",
         "@open-pioneer/notifier": "workspace:^",
-        "@open-pioneer/runtime": "^2.0.2",
+        "@open-pioneer/runtime": "^2.1.0",
         "@open-pioneer/theme": "workspace:^",
         "react": "^18.2.0"
     }
