+++ src/samples/ogc-api-sample/ogc-app/package.json
@@ -4,12 +4,12 @@
     "dependencies": {
         "@open-pioneer/basemap-switcher": "workspace:^",
         "@open-pioneer/chakra-integration": "^1.1.1",
+        "@open-pioneer/http": "^2.1.2",
         "@open-pioneer/map": "workspace:^",
         "@open-pioneer/map-navigation": "workspace:^",
         "@open-pioneer/ogc-features": "workspace:^",
-        "@open-pioneer/http": "^2.1.1",
         "@open-pioneer/react-utils": "workspace:^",
-        "@open-pioneer/runtime": "^2.0.2",
+        "@open-pioneer/runtime": "^2.1.0",
         "@open-pioneer/theme": "workspace:^",
         "@open-pioneer/toc": "workspace:^",
         "ol": "^8.2.0",
