+++ src/samples/test-basemap-switcher/basemap-switcher-app/package.json
@@ -6,7 +6,7 @@
         "@open-pioneer/chakra-integration": "^1.1.1",
         "@open-pioneer/map": "workspace:^",
         "@open-pioneer/react-utils": "workspace:^",
-        "@open-pioneer/runtime": "^2.0.2",
+        "@open-pioneer/runtime": "^2.1.0",
         "@open-pioneer/theme": "workspace:^",
         "react": "^18.2.0"
     },
