+++ src/samples/test-highlight-and-zoom/highlight-and-zoom-app/package.json
@@ -5,7 +5,7 @@
         "@open-pioneer/chakra-integration": "^1.1.1",
         "@open-pioneer/map": "workspace:^",
         "@open-pioneer/react-utils": "workspace:^",
-        "@open-pioneer/runtime": "^2.0.2",
+        "@open-pioneer/runtime": "^2.1.0",
         "react": "^18.2.0"
     },
     "version": "0.0.3"
