+++ src/samples/test-menu-fix/menu-fix-app/package.json
@@ -4,7 +4,7 @@
     "private": true,
     "dependencies": {
         "@open-pioneer/chakra-integration": "^1.1.1",
-        "@open-pioneer/runtime": "^2.0.2",
+        "@open-pioneer/runtime": "^2.1.0",
         "react": "^18.2.0"
     }
 }
